updated model detail naming conentions & moved model detail example to tests
