+++ tests/model_details_example.json
@@ -1,6 +1,6 @@
 {
     "title": "A Generated Model Details",
-    "organization": "Northwestern University",
+    "organization": "Open Source Org",
     "authors": [
         "Matthew Alvarez ", 
         "Jenny Lam", 
@@ -8,10 +8,10 @@
         "Blaine Rothrock"
     ],
     "emails": [
-        "matthewalvarez2021@u.northwestern.edu",
-        "jennylam2021@u.northwestern.edu",
-        "sundar@u.northwestern.edu",
-        "blainerothrock2020@u.northwestern.edu"
+        "matt@email.com",
+        "jenny@email.com",
+        "sundar@email.com",
+        "blaine@email.com"
     ],
     "model_date": "2020-11-15 10:20:56.836960",
     "model_version": "0.0.1",
